+++ MCtrajs/transitionPaths.cpp
@@ -430,7 +430,7 @@ int main(int argc, char **argv)
   vector<double> qfin_av(3, 0);
   double chifin_av = 0.;
   double probafin_av = 0.;
-  long double L_av = 0.;
+  double L_av = 0.;
   int counter = 0;
 
   double ther_meanMeanEner, ther_meanMaxEner, ther_qDistFromStFw, ther_meanJumpsAv, ther_meanJumpsStdDev, ther_meanJumpsMin, ther_meanJumpsMax;
